MVC
Import the Booker MVC on eclipse as an archive file and run it. It will show you take you to the web page.
The CRUD functionality can be assessed from the landing page
VIEW- Click on All Books to access all the books
INSERT- Click on Add Book to add a new book to the database
SEARCH- Enter details of a book(name, or author), and then click on search button. This retrieves information of a book if the book is in the database 
DELETE - Access Delete function from the All Books page and Search Page. click on the Delete icon to delete a book record.
UPDATE - Access Update function from the All Books page and Search Page. click on the Update icon to update a book record.